Extract citation-grounding scorer into common/scorers/citation.py
