+++ src/olmo_eval/common/scorers/__init__.py
@@ -14,6 +14,13 @@ from .base import (
     Scorer,
     SQuADF1Scorer,
 )
+from .citation import (
+    CITATION_GROUP_PROMPT,
+    JUST_HAS_A_TITLE,
+    compute_citation_scores_from_groups,
+    score_citation_group,
+    score_citations_for_sections,
+)
 from .code_execution import CodeExecutionScorer, MultiplEScorer
 from .execution import ContextScorer, ExecutionScorer, SandboxRequiredError
 from .ifeval import IFEvalScorer
@@ -42,7 +49,9 @@ from .trajectory import (
 __all__ = [
     "BitsPerByteScorer",
     "build_openai_judge_fn",
+    "CITATION_GROUP_PROMPT",
     "CodeExecutionScorer",
+    "compute_citation_scores_from_groups",
     "ContextScorer",
     "ExactMatchFlexScorer",
     "ExactMatchScorer",
@@ -50,6 +59,7 @@ __all__ = [
     "F1Scorer",
     "IFEvalScorer",
     "JudgeFn",
+    "JUST_HAS_A_TITLE",
     "LLMJudgeScorer",
     "LogprobScorer",
     "MathVerifyScorer",
@@ -61,6 +71,8 @@ __all__ = [
     "RubricJudgeScorer",
     "SafetyScorer",
     "SandboxRequiredError",
+    "score_citation_group",
+    "score_citations_for_sections",
     "Scorer",
     "SQuADF1Scorer",
     "SimpleQAGrade",
